+++ src/propagators/j2.jl
@@ -285,7 +285,7 @@ end
     ) where {
         Tjd <: Number,
         Tv <: AbstractVector
-    } -> KeplerianElements{MeanAnomaly, Float64, Float64}, SMatrix{6, 6, Float64}
+    } -> KeplerianElements{MeanAnomaly, Float64, T}, SMatrix{6, 6, T}
 
 Fit a set of mean Keplerian elements for the J2 orbit propagator using the osculating
 elements represented by a set of position vectors `vr_i` [m] and a set of velocity vectors
@@ -294,12 +294,15 @@ elements represented by a set of position vectors `vr_i` [m] and a set of veloci
 
 !!! note
 
-    This algorithm version will allocate a new J2 propagator with the default constants
-    `J2C_EGM2008`. If another set of constants are required, use the function
-    [`fit_j2_mean_elements!`](@ref) instead.
+    This algorithm version will allocate a new J2 propagator with the constants `j2c`. If
+    the allocation must be avoided, use the function [`fit_j2_mean_elements!`](@ref)
+    instead.
 
 # Keywords
 
+- `j2c::J2PropagatorConstants{T}`: J2 orbit propagator constants (see
+    [`J2PropagatorConstants`](@ref)), whose number type `T` is used in the fitting.
+    (**Default**: `J2C_EGM2008`)
 - `atol::Number`: Tolerance for the residual absolute value. If the residual is lower than
     `atol` at any iteration, the computation loop stops.
     (**Default**: 2e-4)
@@ -337,8 +340,8 @@ elements represented by a set of position vectors `vr_i` [m] and a set of veloci
 
 # Returns
 
-- `KeplerianElements{MeanAnomaly, Float64, Float64}`: Fitted Keplerian elements.
-- `SMatrix{6, 6, Float64}`: Final covariance matrix of the least-square algorithm.
+- `KeplerianElements{MeanAnomaly, Float64, T}`: Fitted Keplerian elements.
+- `SMatrix{6, 6, T}`: Final covariance matrix of the least-square algorithm.
 
 # Examples
 
@@ -390,13 +393,16 @@ KeplerianElements{MeanAnomaly, Float64, Float64}:
 ```
 """
 function fit_j2_mean_elements(
-    vjd::AbstractVector{Tjd}, vr_i::AbstractVector{Tv}, vv_i::AbstractVector{Tv}; kwargs...
-) where {Tjd <: Number, Tv <: AbstractVector}
-    # Allocate the J2 propagator structure that will propagate the mean elements.
-    j2d = J2Propagator{Float64, Float64}()
-
-    # Assign the constants, which are used in the initialization.
-    j2d.j2c = J2C_EGM2008
+    vjd::AbstractVector{Tjd},
+    vr_i::AbstractVector{Tv},
+    vv_i::AbstractVector{Tv};
+    j2c::J2PropagatorConstants{T} = J2C_EGM2008,
+    kwargs...,
+) where {Tjd <: Number, Tv <: AbstractVector, T <: Number}
+    # Allocate the J2 propagator structure that will propagate the mean elements and assign
+    # the constants, which are used in the initialization.
+    j2d = J2Propagator{Float64, T}()
+    j2d.j2c = j2c
 
     return fit_j2_mean_elements!(j2d, vjd, vr_i, vv_i; kwargs...)
 end
@@ -532,7 +538,8 @@ end
 """
     update_j2_mean_elements_epoch(
         orb::KeplerianElements,
-        new_epoch::Union{Number, DateTime}
+        new_epoch::Union{Number, DateTime};
+        kwargs...
     ) -> KeplerianElements{MeanAnomaly}
 
 Update the epoch of the mean elements `orb` using a J2 orbit propagator to `new_epoch`,
@@ -540,9 +547,15 @@ which can be represented by a Julian Day or a `DateTime`.
 
 !!! note
 
-    This algorithm version will allocate a new J2 propagator with the default constants
-    `J2C_EGM2008`. If another set of constants are required, use the function
+    This algorithm version will allocate a new J2 propagator with the constants `j2c`. If
+    the allocation must be avoided, use the function
     [`update_j2_mean_elements_epoch!`](@ref) instead.
+
+# Keywords
+
+- `j2c::J2PropagatorConstants`: J2 orbit propagator constants (see
+    [`J2PropagatorConstants`](@ref)), converted to the element type of `orb`.
+    (**Default**: `J2C_EGM2008`)
 
 # Examples
 
@@ -577,13 +590,14 @@ KeplerianElements{MeanAnomaly, Float64, Float64}:
 ```
 """
 function update_j2_mean_elements_epoch(
-    orb::KeplerianElements{Tanomaly, Tepoch, T}, new_epoch::Union{Number, DateTime}
+    orb::KeplerianElements{Tanomaly, Tepoch, T},
+    new_epoch::Union{Number, DateTime};
+    j2c::J2PropagatorConstants = J2C_EGM2008,
 ) where {Tanomaly <: AbstractAnomaly, Tepoch <: Number, T <: Number}
-    # Allocate the J2 propagator structure that will propagate the mean elements.
+    # Allocate the J2 propagator structure that will propagate the mean elements and assign
+    # the constants, which are used in the initialization.
     j2d = J2Propagator{Tepoch, T}()
-
-    # Assign the constants, which are used in the initialization.
-    j2d.j2c = J2C_EGM2008
+    j2d.j2c = convert(J2PropagatorConstants{T}, j2c)
 
     return update_j2_mean_elements_epoch!(j2d, orb, new_epoch)
 end