+++ src/propagators/j2osc.jl
@@ -183,7 +183,7 @@ end
     ) where {
         Tjd <: Number,
         Tv <: AbstractVector
-    } -> KeplerianElements{MeanAnomaly, Float64, Float64}, SMatrix{6, 6, Float64}
+    } -> KeplerianElements{MeanAnomaly, Float64, T}, SMatrix{6, 6, T}
 
 Fit a set of mean Keplerian elements for the J2 osculating orbit propagator using the
 osculating elements represented by a set of position vectors `vr_i` [m] and a set of
@@ -192,12 +192,15 @@ the array `vjd` [Julian Day].
 
 !!! note
 
-    This algorithm version will allocate a new J2 osculating propagator with the default
-    constants `J2C_EGM2008`. If another set of constants are required, use the function
+    This algorithm version will allocate a new J2 osculating propagator with the constants
+    `j2c`. If the allocation must be avoided, use the function
     [`fit_j2osc_mean_elements!`](@ref) instead.
 
 # Keywords
 
+- `j2c::J2PropagatorConstants{T}`: J2 orbit propagator constants (see
+    [`J2PropagatorConstants`](@ref)), whose number type `T` is used in the fitting.
+    (**Default**: `J2C_EGM2008`)
 - `atol::Number`: Tolerance for the residual absolute value. If the residual is lower than
     `atol` at any iteration, the computation loop stops.
     (**Default**: 2e-4)
@@ -235,8 +238,8 @@ the array `vjd` [Julian Day].
 
 # Returns
 
-- `KeplerianElements{MeanAnomaly, Float64, Float64}`: Fitted Keplerian elements.
-- `SMatrix{6, 6, Float64}`: Final covariance matrix of the least-square algorithm.
+- `KeplerianElements{MeanAnomaly, Float64, T}`: Fitted Keplerian elements.
+- `SMatrix{6, 6, T}`: Final covariance matrix of the least-square algorithm.
 
 # Examples
 
@@ -276,16 +279,19 @@ KeplerianElements{MeanAnomaly, Float64, Float64}:
 ```
 """
 function fit_j2osc_mean_elements(
-    vjd::AbstractVector{Tjd}, vr_i::AbstractVector{Tv}, vv_i::AbstractVector{Tv}; kwargs...
-) where {Tjd <: Number, Tv <: AbstractVector}
-    # Allocate the J2 propagator structure that will propagate the mean elements.
-    j2d = J2Propagator{Float64, Float64}()
-
-    # Assign the constants, which are used in the initialization.
-    j2d.j2c = J2C_EGM2008
+    vjd::AbstractVector{Tjd},
+    vr_i::AbstractVector{Tv},
+    vv_i::AbstractVector{Tv};
+    j2c::J2PropagatorConstants{T} = J2C_EGM2008,
+    kwargs...,
+) where {Tjd <: Number, Tv <: AbstractVector, T <: Number}
+    # Allocate the J2 propagator structure that will propagate the mean elements and assign
+    # the constants, which are used in the initialization.
+    j2d = J2Propagator{Float64, T}()
+    j2d.j2c = j2c
 
     # Allocate the J2 osculating propagator structure.
-    j2oscd = J2OsculatingPropagator{Float64, Float64}()
+    j2oscd = J2OsculatingPropagator{Float64, T}()
     j2oscd.j2d = j2d
 
     return fit_j2osc_mean_elements!(j2oscd, vjd, vr_i, vv_i; kwargs...)
@@ -410,7 +416,8 @@ end
 """
     update_j2osc_mean_elements_epoch(
         orb::KeplerianElements,
-        new_epoch::Union{Number, DateTime}
+        new_epoch::Union{Number, DateTime};
+        kwargs...
     ) -> KeplerianElements{MeanAnomaly}
 
 Update the epoch of the mean elements `orb` using a J2 osculating orbit propagator to
@@ -418,9 +425,15 @@ Update the epoch of the mean elements `orb` using a J2 osculating orbit propagat
 
 !!! note
 
-    This algorithm version will allocate a new J2 osculating propagator with the default
-    constants `J2C_EGM2008`. If another set of constants are required, use the function
+    This algorithm version will allocate a new J2 osculating propagator with the constants
+    `j2c`. If the allocation must be avoided, use the function
     [`update_j2osc_mean_elements_epoch!`](@ref) instead.
+
+# Keywords
+
+- `j2c::J2PropagatorConstants`: J2 orbit propagator constants (see
+    [`J2PropagatorConstants`](@ref)), converted to the element type of `orb`.
+    (**Default**: `J2C_EGM2008`)
 
 # Examples
 
@@ -455,13 +468,14 @@ KeplerianElements{MeanAnomaly, Float64, Float64}:
 ```
 """
 function update_j2osc_mean_elements_epoch(
-    orb::KeplerianElements{Tanomaly, Tepoch, T}, new_epoch::Union{Number, DateTime}
+    orb::KeplerianElements{Tanomaly, Tepoch, T},
+    new_epoch::Union{Number, DateTime};
+    j2c::J2PropagatorConstants = J2C_EGM2008,
 ) where {Tanomaly <: AbstractAnomaly, Tepoch <: Number, T <: Number}
-    # Allocate the J2 propagator structure that will propagate the mean elements.
+    # Allocate the J2 propagator structure that will propagate the mean elements and assign
+    # the constants, which are used in the initialization.
     j2d = J2Propagator{Tepoch, T}()
-
-    # Assign the constants, which are used in the initialization.
-    j2d.j2c = J2C_EGM2008
+    j2d.j2c = convert(J2PropagatorConstants{T}, j2c)
 
     # Allocate the J2 osculating propagator structure.
     j2oscd = J2OsculatingPropagator{Tepoch, T}()